Move artwork components to Components
