+++ src/Styleguide/Pages/Artwork/Components/ArtistInfo.tsx
