+++ src/Styleguide/Pages/Artwork/Components/Banner.tsx
