+++ src/Styleguide/Pages/Artwork/Components/Gallery.tsx
