+++ src/Styleguide/Pages/Artwork/Components/ImageCarousel.tsx
