+++ src/Styleguide/Pages/Artwork/Components/Sidebar.tsx
@@ -2,13 +2,13 @@ import { Serif } from "@artsy/palette"
 import React, { Component } from "react"
 import styled from "styled-components"
 import { Separator } from "Styleguide/Elements/Separator"
-import { ArtworkMetadata } from "Styleguide/Pages/Artwork/Sidebar/ArtworkMetadata"
-import { AuctionPartnerInfo } from "Styleguide/Pages/Artwork/Sidebar/AuctionPartnerInfo"
-import { Commercial } from "Styleguide/Pages/Artwork/Sidebar/Commercial"
-import { CurrentBidInfo } from "Styleguide/Pages/Artwork/Sidebar/CurrentBidInfo"
-import { ExtraLinks } from "Styleguide/Pages/Artwork/Sidebar/ExtraLinks"
-import { PartnerInfo } from "Styleguide/Pages/Artwork/Sidebar/PartnerInfo"
 import { Artists } from "./Sidebar/Artists"
+import { ArtworkMetadata } from "./Sidebar/ArtworkMetadata"
+import { AuctionPartnerInfo } from "./Sidebar/AuctionPartnerInfo"
+import { Commercial } from "./Sidebar/Commercial"
+import { CurrentBidInfo } from "./Sidebar/CurrentBidInfo"
+import { ExtraLinks } from "./Sidebar/ExtraLinks"
+import { PartnerInfo } from "./Sidebar/PartnerInfo"
 
 export interface ArtworkSidebarProps {
   readonly artwork: {
