+++ src/Styleguide/Pages/Artwork/Components/Sidebar/Artists.tsx
