+++ src/Styleguide/Pages/Artwork/Components/Sidebar/CurrentBidInfo.tsx
