+++ src/Styleguide/Pages/Artwork/Components/Sidebar/PartnerInfo.tsx
