+++ src/Styleguide/Pages/Artwork/Components/__stories__/ArtistInfo.story.tsx
