+++ src/Styleguide/Pages/Artwork/Components/__stories__/Banner.story.tsx
