+++ src/Styleguide/Pages/Artwork/Components/__stories__/Gallery.story.tsx
