+++ src/Styleguide/Pages/Artwork/Components/__stories__/ImageCarousel.story.tsx
