+++ src/Styleguide/Pages/Artwork/Components/__stories__/OtherAuctions.story.tsx
