+++ src/Styleguide/Pages/Artwork/Components/__stories__/Sidebar/Commercial.story.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import { storiesOf } from "storybook/storiesOf"
-import { Commercial } from "Styleguide/Pages/Artwork/Sidebar/Commercial"
+import { Commercial } from "Styleguide/Pages/Artwork/Components/Sidebar/Commercial"
 import {
   ContactForPriceWork,
   FoSaleArtworkNoEditions,
