+++ src/Styleguide/Pages/Artwork/index.tsx
@@ -3,20 +3,20 @@ import styled from "styled-components"
 import { FullArtworkGrid } from "Styleguide/Components/ArtworkGrid"
 import { Tab, Tabs } from "Styleguide/Components/Tabs"
 import { Col, Row } from "Styleguide/Elements/Grid"
-import { OtherWorks } from "Styleguide/Pages/Artwork/OtherWorks"
 import { Boot } from "Styleguide/Pages/Boot"
 import { exhibitions } from "Styleguide/Pages/Fixtures//SelectedExhibitions"
 import { bio } from "Styleguide/Pages/Fixtures/ArtistBio"
 import { imageData } from "Styleguide/Pages/Fixtures/Artwork/imageData"
 import { RegularArtwork2Editions } from "Styleguide/Pages/Fixtures/Artworks"
 import { insights } from "Styleguide/Pages/Fixtures/MarketInsights"
-import { AboutTheWork } from "./AboutTheWork"
-import { ArtistInfo } from "./ArtistInfo"
-import { Banner } from "./Banner"
-import { Bibliography } from "./Bibliography"
-import { ExhibitionHistory } from "./ExhibitionHistory"
-import { ImageCarousel } from "./ImageCarousel"
-import { Sidebar } from "./Sidebar"
+import { AboutTheWork } from "./Components/AboutTheWork"
+import { ArtistInfo } from "./Components/ArtistInfo"
+import { Banner } from "./Components/Banner"
+import { Bibliography } from "./Components/Bibliography"
+import { ExhibitionHistory } from "./Components/ExhibitionHistory"
+import { ImageCarousel } from "./Components/ImageCarousel"
+import { OtherWorks } from "./Components/OtherWorks"
+import { Sidebar } from "./Components/Sidebar"
 
 export class Artwork extends Component {
   state = {
