+++ src/__stories__/storiesOf.js
@@ -1,21 +1,9 @@
 import React from "react"
 import { storiesOf as _storiesOf } from "@storybook/react"
-import { Theme, themeProps } from "@artsy/palette"
-import { GlobalStyles } from "../Styleguide/Elements/GlobalStyles"
-import { ResponsiveProvider } from "../Styleguide/Utils/Responsive"
-
-const breakpoints = themeProps.mediaQueries
+import { Boot } from "../Styleguide/Pages/Boot"
 
 export function storiesOf(desc, mod) {
   return _storiesOf(desc, mod).addDecorator(storyFn => {
-    return (
-      <Theme>
-        <GlobalStyles>
-          <ResponsiveProvider breakpoints={breakpoints}>
-            {storyFn()}
-          </ResponsiveProvider>
-        </GlobalStyles>
-      </Theme>
-    )
+    return <Boot>{storyFn()}</Boot>
   })
 }
